Move merge_files to new module traceutils
