+++ reprozip/reprozip/pack.py
@@ -24,7 +24,7 @@ from reprozip import __version__ as reprozip_version
 from reprozip.common import File, load_config, save_config, \
     record_usage_package
 from reprozip.tracer.linux_pkgs import identify_packages
-from reprozip.tracer.trace import merge_files
+from reprozip.traceutils import merge_files
 from reprozip.utils import iteritems
