+++ reprozip/reprozip/traceutils.py
@@ -1,0 +1,36 @@
+# Copyright (C) 2014-2016 New York University
+# This file is part of ReproZip which is released under the Revised BSD License
+# See file LICENSE for full license details.
+
+"""Additional manipulations for traces.
+
+These are operations on traces that are not directly related to the tracing
+process itself.
+"""
+
+from reprozip.tracer.trace import TracedFile
+from reprozip.utils import listvalues
+
+
+def merge_files(newfiles, newpackages, oldfiles, oldpackages):
+    """Merges two sets of packages and files.
+    """
+    files = set(oldfiles)
+    files.update(newfiles)
+
+    packages = dict((pkg.name, pkg) for pkg in newpackages)
+    for oldpkg in oldpackages:
+        if oldpkg.name in packages:
+            pkg = packages[oldpkg.name]
+            # Here we build TracedFiles from the Files so that the comment
+            # (size, etc) gets set
+            s = set(TracedFile(fi.path) for fi in oldpkg.files)
+            s.update(pkg.files)
+            oldpkg.files = list(s)
+            packages[oldpkg.name] = oldpkg
+        else:
+            oldpkg.files = [TracedFile(fi.path) for fi in oldpkg.files]
+            packages[oldpkg.name] = oldpkg
+    packages = listvalues(packages)
+
+    return files, packages
